update code v1 payment

+++ Payment.App/Data/EntityConfigurations/VnPayTransEntityTypeConfiguration.cs
@@ -6,7 +6,7 @@ public class VnPayTransEntityTypeConfiguration : IEntityTypeConfiguration<VnPayT
     {
         builder.ToTable("VNPayTrans");
         builder.HasKey(p => p.Id);
-        builder.Property(p => p.Id).HasColumnType("INT");
+        builder.Property(p => p.Id).HasColumnType("BIGINT");
         builder.Property(p => p.TxnId).HasColumnType("NVARCHAR(100)");
         builder.Property(p => p.State).HasColumnType("INT");
         builder.Property(p => p.Amount).HasColumnType("NVARCHAR(100)");
